pyproject.toml
[project]
name = "agenda-atronea"
version = "0.1.0"
description = ""
authors = [
    {name = "lautaropacella"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.48.1,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "streamlit-authenticator (>=0.4.2,<0.5.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "plotly (>=6.3.0,<7.0.0)",
    "pydrive2 (>=1.21.3,<2.0.0)",
    "pyopenssl (<=24.2.1)",
    "cryptography (<44)",
    "streamlit-oauth (>=0.1.14,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false